decouple use appModel
